refactor(auth): restructure DTO package hierarchy
Relocate `dao.dto` classes to `dto` package to align with project structure and improve consistency. Update imports across the project to reflect the new structure.
